+++ src/commands/economy/index.js
@@ -3,6 +3,7 @@
  * Copyright (c) 2021 The Welcome-Bot Team and Contributors
  * Licensed under Lesser General Public License v2.1 (LGPl-2.1 - https://opensource.org/licenses/lgpl-2.1.php)
  */
+const emojis = require("../../data/customEmojis.json");
 const commands = require("fs")
     .readdirSync(__dirname)
     .filter((file) => file !== "index.js" && file.endsWith(".js"))
@@ -11,8 +12,8 @@ const commands = require("fs")
 module.exports = {
     commands,
     metadata: {
-        name: "Fun",
-        key: "fun",
-        emoji: "<:fun:854002049095303188>",
+        name: "Economy",
+        key: "economy",
+        emoji: emojis.economy,
     },
 };