+++ CHANGELOG.md
@@ -16,6 +16,15 @@ The displayed date is in the format `DD-MM-YYYY`
 
 - New cmd `emojify`
 - refactor(disable cmd): Disable commands using disable cmd now!
+- feat: New cmd `daily` and new category `Economy`
+- feat: New cmd `balance`
+- feat: New cmd `deposit`
+- feat: New cmd `give`
+- feat: New cmd `beg`
+- feat: New cmd `withdraw`
+- feat: New cmd `profile`
+- feat: New cmd `setbio`
+- feat: New cmd `rob`
 
 ### Changes